requirements.txt
Django==4.2.4
django-debug-toolbar==4.1.0
djangorestframework==3.14.0